Go back to JSON parsing
Swift is bad at compiling big collection literals. This was made faster by
splitting the literal into .append calls but it still was pretty slow.

Combine that with the fact that the entire thing just crashed on watchOS
and JSON just became the better option.

The biggest drawback with the JSON way is that we need to assume parsing
and loading from disk works. Or compromise on the API/developer
experience. I chose for the “we'll just assume it works” approach
because most of the time it will.