Always look up emojis by their unified string
Works around issues where certain emojis contain additional unicode
scalars that aren't in our lookup table.

For example ☕️ is comprised of U+2615-U+FE0F, but is just `2615` in our
lookup table because U+FE0F (the emoji variation selector) doesn't do anything when
there is no text representation of the emoji. However it is required
for SNOWMAN because it has a text representation (☃).

Because some emoji characters can be represented with and without the
variation selector we look up both variants in our lookup table. For
example with the HOT BEVERAGE character we look up `2615-FE0F` first
then just `2615` because the first lookup returned `nil`.

Conversely with the SNOWMAN character we lookup `2603-FE0F` first and
immediately find it because `2603` isn't a emoji.

See https://symbl.cc/en/FE0F/ for more information about the variation
selector.